requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
attrs==25.3.0
Authlib
banks==2.2.0
beautifulsoup4==4.13.5
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
exceptiongroup==1.3.0
filelock==3.19.1
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
griffe==1.13.0
groq==0.31.0
grpcio==1.74.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.13.3
llama-index-cli==0.5.0
llama-index-core==0.13.3
llama-index-embeddings-huggingface==0.6.0
llama-index-embeddings-openai==0.5.0
llama-index-indices-managed-llama-cloud==0.9.2
llama-index-instrumentation==0.4.0
llama-index-llms-groq==0.4.0
llama-index-llms-openai==0.5.4
llama-index-llms-openai-like==0.5.0
llama-index-readers-file==0.5.2
llama-index-readers-llama-parse==0.5.0
llama-index-vector-stores-qdrant==0.8.2
llama-index-workflows==1.3.0
llama-parse==0.6.54
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
narwhals==2.2.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.6
openai==1.102.0
packaging==25.0
pandas==2.2.3
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
platformdirs==4.4.0
portalocker==3.2.0
propcache==0.3.2
protobuf==6.32.0
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
PyMuPDF==1.26.4
pypdf==6.0.0
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
qdrant-client==1.15.1
referencing==0.36.2
regex==2025.8.29
requests==2.32.5
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
sentence-transformers==5.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
streamlit==1.49.1
streamlit-authenticator
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.0
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.56.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wrapt==1.17.3
yarl==1.20.1
